(#3577) - edit intro docs
The "Learn" page is probably the first place where
users will land after the home page, and I wanted to
clean it up and update the description a bit.

As for the API, I cleaned up some small inconsistencies
and added a longer intro section on promises/callbacks.
At some point, I would like to go in there and just remove
the callback examples entirely. I think we should deprecate
the callback API.